WORKIN ON DISPLAY METHODS

+++ dotNet5782_4199_5185/BL/DroneInCharging.cs
@@ -9,7 +9,7 @@ namespace IBL.BO
     public class DroneInCharging
     {
         public int Id {get;set; }
-        public int Battary {get;set; }
+        public double Battary {get;set; }
 
         public override string ToString()
         {
